+++ CHANGELOG.md
@@ -18,6 +18,22 @@
 - Test for `order > 0` in `bellman_ford_moore` and other algorithms that take a digraph.
 - Replace the blanket implementations in `op` with custom implementations for the types in `repr`. In the documentation, describe examples of both the direct and the trait-based implementation of the traits.
 
+## [0.100.0] - 2024-10-04
+
+Added
+
+- Add benchmark `algo_bellman_ford_moore_new`.
+- Add benchmark `op_is_complete`.
+
+Changed
+
+- Breaking: replace `impl<D> IsComplete for D where D: HasEdge + Order` with `impl IsComplete for [AdjacencyList, AdjacencyMap, AdjacencyMatrix, EdgeList]`.
+- Speed up `AdjacencyList::is_complete`.
+- Speed up `AdjacencyMap::is_complete`.
+- Speed up `AdjacencyMatrix::is_complete`.
+- Speed up `BellmanFordMoore::new`.
+- Speed up `EdgeList::is_complete`.
+
 ## [0.99.0] - 2024-10-04
 
 Added